Moving aiotarstream to the deployment module
Tar streaming logic is only used in the `deployment` package, and
cross-referencing it from the `data` package can generate errors due to
circular dependencies at import time. Moving the logic directly into the
deployment packages solves these issues and makes modules more self contained.
