+++ streamflow/deployment/connector/ssh.py
@@ -20,8 +20,8 @@ from streamflow.core.data import StreamWrapperContext
 from streamflow.core.deployment import Connector, Location
 from streamflow.core.exception import WorkflowExecutionException
 from streamflow.core.scheduling import AvailableLocation, Hardware
-from streamflow.data import aiotarstream
-from streamflow.data.stream import StreamReaderWrapper, StreamWriterWrapper
+from streamflow.deployment import aiotarstream
+from streamflow.deployment.stream import StreamReaderWrapper, StreamWriterWrapper
 from streamflow.deployment.connector.base import BaseConnector, extract_tar_stream
 from streamflow.log_handler import logger
